+++ src/sphinxawesome_theme/html_translator.py
@@ -12,50 +12,28 @@ from typing import Any, Dict
 
 from docutils import nodes
 from docutils.nodes import Element, Text
+from sphinx import addnodes
 from sphinx.application import Sphinx
 from sphinx.locale import _
 from sphinx.util import logging
 from sphinx.writers.html5 import HTML5Translator
 
 from . import __version__
+from .icons import ICONS
 
 logger = logging.getLogger(__name__)
-
-# icons are from Material Design icon set
-ICONS = {
-    # https://material.io/resources/icons/?icon=content_copy
-    "copy": (
-        "<svg xmlns='http://www.w3.org/2000/xvg' viewBox='0 0 24 24' "
-        "fill=currentColor aria-hidden='true'>"
-        "<path d='M16 1H4c-1.1 0-2 .9-2 2v14h2V3h12V1zm3 "
-        "4H8c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h11c1.1 "
-        "0 2-.9 2-2V7c0-1.1-.9-2-2-2zm0 16H8V7h11v14z' /></svg>"
-    ),
-    # https://material.io/resources/icons/?icon=link
-    "headerlink": (
-        '<svg xmlns="http://www.w3.org/2000/svg" '
-        'pointer-events="none" viewBox="0 0 24 24">'
-        '<path d="M3.9 12c0-1.71 1.39-3.1 '
-        "3.1-3.1h4V7H7c-2.76 0-5 2.24-5 5s2.24 "
-        "5 5 5h4v-1.9H7c-1.71 0-3.1-1.39-3.1-3.1zM8 "
-        "13h8v-2H8v2zm9-6h-4v1.9h4c1.71 0 3.1 1.39 3.1 "
-        "3.1s-1.39 3.1-3.1 3.1h-4V17h4c2.76 0 5-2.24 "
-        '5-5s-2.24-5-5-5z"/></svg>'
-    ),
-    # https://material.io/resources/icons/?icon=open_in_new
-    "external_link": (
-        '<svg xmlns="http://www.w3.org/2000/svg" '
-        'viewBox="0 0 24 24" class="external-link-icon" aria-hidden="true">'
-        '<path d="M19 19H5V5h7V3H5a2 2 0 00-2 2v14a2 2 0 '
-        "002 2h14c1.1 0 2-.9 2-2v-7h-2v7zM14 "
-        '3v2h3.59l-9.83 9.83 1.41 1.41L19 6.41V10h2V3h-7z"/></svg>'
-    ),
-}
 
 COPY_BUTTON = (
     "<button class='copy tooltipped tooltipped-nw' aria-label='Copy this code'>"
     + ICONS["copy"]
     + "</button>\n"
+)
+
+EXPAND_MORE_BUTTON = (
+    "<button class='expand-more tooltipped tooltipped-nw' "
+    "aria-label='Expand this section' aria-expanded='false'>"
+    + ICONS["expand_more"]
+    + "</button>"
 )
 
 
@@ -176,6 +154,20 @@ class AwesomeHTMLTranslator(HTML5Translator):
         else:
             self.body.append("</p>\n")
 
+    def visit_desc_signature(self, node: Element) -> None:
+        """Add the accordion class to the <dt> element.
+
+        This will make the definition list collapsible,
+        if the configuration option ``html_collapsible_definitions``
+        is set.
+        """
+        # only add this, if the following dd is not empty.
+        dd = node.next_node(addnodes.desc_content, siblings=True)
+        if self.config.html_collapsible_definitions and len(dd.astext()) > 0:
+            self.body.append(self.starttag(node, "dt", CLASS="accordion"))
+        else:
+            self.body.append(self.starttag(node, "dt"))
+
     def depart_desc_signature(self, node: Element) -> None:
         """Change permalinks for code definitions.
 
@@ -184,13 +176,25 @@ class AwesomeHTMLTranslator(HTML5Translator):
         """
         if not node.get("is_multiline"):
             self.add_permalink_ref(node, _("Copy link to this definition."))
+        dd = node.next_node(addnodes.desc_content, siblings=True)
+        if self.config.html_collapsible_definitions and len(dd.astext()) > 0:
+            self.body.append(EXPAND_MORE_BUTTON)
         self.body.append("</dt>\n")
 
     def depart_desc_signature_line(self, node: Element) -> None:
         """Change permalinks for code definitions."""
         if node.get("add_permalink"):
             self.add_permalink_ref(node.parent, _("Copy link to this definition."))
+        if self.config.html_collapsible_definitions:
+            self.body.append(EXPAND_MORE_BUTTON)
         self.body.append("<br />")
+
+    def visit_desc_content(self, node: Element) -> None:
+        """Add panel class to definitions."""
+        if self.config.html_collapsible_definitions and len(node.astext()) > 0:
+            self.body.append(self.starttag(node, "dd", CLASS="panel"))
+        else:
+            self.body.append(self.starttag(node, "dd"))
 
     def visit_section(self, node: Element) -> None:
         """Use semantic <section> elements."""
@@ -325,6 +329,7 @@ def setup(app: "Sphinx") -> Dict[str, Any]:
     """
     app.set_translator("html", AwesomeHTMLTranslator)
     app.set_translator("dirhtml", AwesomeHTMLTranslator)
+    app.add_config_value("html_collapsible_definitions", False, "html")
 
     return {
         "version": __version__,