+++ src/theme-src/app.js
@@ -4,6 +4,7 @@ import { navFunctions, collapsibleNav } from "./js/nav";
 import { searchPane, clearSearchHighlights, searchEvents } from "./js/search";
 import { scrollActive } from "./js/scroll";
 import { copyEvents } from "./js/copy";
+import { collapsible } from "./js/collapsible";
 
 navFunctions();
 searchPane();
@@ -12,3 +13,4 @@ searchEvents();
 copyEvents();
 collapsibleNav();
 scrollActive();
+collapsible();