+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+MASTER
+~~~~~~
+
+- feat: new option to make autodocs definitions collapsible (PR #167)
+
 1.15.1
 ~~~~~~